Fix return values of `and` and `or`
Builtin `and` always returns the result of the last expression it
evaluated.

A code comment for `and` was stupidly referencing the common lisp
standard when common lisp has no symbol to represent boolean "false".

We want logical operators with arguments that evaluate to boolean values
to return boolean values themselves in order to serialize json properly.

Because of the above, the operator `or` has also been changed to return
`false` when called without arguments.