New LType, LTaggedVal, a container for user-defined data types
LTaggedVal is an arbitrary lisp value that has been "tagged" with a
custom type name (symbol).  That is, an LTaggedVal is an instance of a
user-defined type and the value contained within is the instance data.
The following functions and macros have been defined for working with
LTaggedVal values:

- ``deftype''
  Define a new type -- binding its name to a typedef
- ``new''
  Instantiate a tagged-value from a typedef
- ``type''
  Returns a symbol representing the type of its argument (works for
  any value)
- ``type?''
  Returns a boolean -- true iff the given type specifier matches the
  type of the given object
- ``tagged-value?''
  Returns a bool -- true iff the given object is a tagged-value
- ``user-data''
  Unwraps tagged-value and returns the underlying value

Examples of user-defined types: _examples/random/oop.lisp

libschema is also updated to support tagged-values

- `s:tagged-value` type-level type constraint for tagged-values.

- `s:make-validator` function supports tagged-values and returns
  function usable by `s:validate` for schema validation.